refactor: generalize hardwareDetailsView function
Closes #767
