+++ backend/kernelCI_app/constants/hardwareDetails.py
@@ -1,0 +1,2 @@
+SELECTED_HEAD_TREE_VALUE = "head"
+STATUS_FAILED_VALUE = "FAIL"
